requirements.txt
ccxt==4.2.25
aiohttp==3.9.1
asyncio==3.4.3
python-dotenv==1.0.0
PyYAML==6.0.1
pandas==2.1.4
numpy==1.26.2
websockets==12.0
requests==2.31.0
web3==6.15.1
pytest==7.4.4
colorama==0.4.6
tabulate==0.9.0
schedule==1.2.0